Enable extra compiler warnings & fix

+++ DigiMeSDK/Utility/NSData+DMECrypto.m
@@ -22,7 +22,7 @@
     NSUInteger          dataLength  = [self length];
     NSMutableString     *hexString  = [NSMutableString stringWithCapacity:(dataLength * 2)];
     
-    for (int i = 0; i < dataLength; ++i)
+    for (NSUInteger i = 0; i < dataLength; ++i)
         [hexString appendString:[NSString stringWithFormat:@"%02lx", (unsigned long)dataBuffer[i]]];
     
     return [NSString stringWithString:hexString];
